+++ src/models/ZoneStatus.ts
@@ -9,12 +9,14 @@ export class ZoneStatus {
   private readonly _latchedAlarmState: number;
   private readonly _armingState: number;
   private readonly _troubleAcknowledged: boolean;
+  private readonly _bypassed: boolean;
 
   constructor(status: number) {
     this._currentState = status & 0b00000011;
     this._latchedAlarmState = status & 0b00001100;
     this._armingState = status & 0b00110000;
     this._troubleAcknowledged = (status & 0b01000000) === 0b01000000;
+    this._bypassed = (status & 0b00100000) === 0b00100000;
   }
 
   get currentState(): ZoneStates {
@@ -39,5 +41,9 @@ export class ZoneStatus {
 
   get troubleAcknowledged(): boolean {
     return this._troubleAcknowledged;
+  }
+
+  get bypassed(): boolean {
+    return this._bypassed;
   }
 }