+++ CHANGELOG.md
@@ -8,6 +8,7 @@ All notable changes to this project will be documented in this file.
 * Support platformAccessory Identify event
 * Cache discovered area, zone, unit, button, thermostat & code IDs for faster startup
 * [FIX] Pushover notification may cause plugin to crash when internet connection is down
+* Switch accessories for Bypass Zones
 
 ## 1.3.3 (2020-12-30)